requirements.txt
Django~=4.2
djangorestframework~=3.16.0
chardet~=5.2.0